Calinski-Harabasz Index (CHI / ICH)

Methods to determine the value of the Calinski-Harabasz Index of a partition with elements variable(s) compatible with euclidean distance. This index is defined by between variance divided by within variance, each divided by their degree of liberty. Depending of valueOnly, the value returned by calinski_harabasz if only the index or a list with more data. It can be useful for instance to update the index after some change instead of start from the beginning.

Usage

calinski_harabasz(data, partition, valueOnly = TRUE)

Arguments

data

A data frame with one row per element, containing context. Each variable included must be compatible with the euclidean distance (and so have to be quantitative) because CHI use the variance.

partition

The partition of the problem: a vector with a length equal to the number of rows of data.

valueOnly

TRUE if the only value needed is the index. FALSE if the user wants more information.

Value

Depends on valueOnly. If valueOnly = TRUE, returns only the Calinski-Harabasz index (positive double value). If valueOnly = FALSE, returns a list containing the following elements:

"k": number of clusters in the partition (strictly positive integer)

"nbElementsClusters": strictly positive integers vector of size k giving the number of elements of each cluster.

"withinClusters": A positive double vector of size k containing the within sum of squares of each cluster.

"W": The within sum of squares of the partition, i.e. the sum of the empirical cluster variances weighted by the number of elements of each clusters. (positive real value)

"betweenClusters": A positive double vector of size k containing for each cluster the square euclidean norm between their and the global centroid (i.e. ||c_i - c||²)

"B" : The between sum of squares of the partition, i.e. the sum of the square euclidean norm given in "betweenClusters" weighted by the number of elements of each cluster.

"CHI" : The Calinski-Harabasz Index (positive real value)

Details

Some special cases can be considered:

n = 1 : their is only one possible partition so it is optimal by definition. Hence we take CHI = Inf.

k = 1: the unique cluster of the partition is the entire set. Here the between sum of squares (B) is considered equal to 0. In the index formula B is divided by k-1 which equals 0 here. Except when n = 1 this partition is of no interest so we assign CHI = 0.

k = n : each cluster is made of one element. Consequently the empirical biased variance of each cluster is equal to 0, so is W. In the index formula W is divided by n - k which is equal to 0 here. In that case compactness is optimal but separation is at its worth. The index is set to NaN.

References

Calinski, T., and Harabasz, J. (1974) A Dendrite Method for Cluster Analysis, Communications in Statistics, 3, 1-27.

See also