Find the nearest neighbor of some points

For a set of points (subsetPoints), find the nearest neighbor in another set (subsetNeighbors). Distances are given in a p' x n' matrix (distances), with p' >= p (p the number of points) and n' >= n (n the number of neighbors). If points and neighbors come from the same set, put inner to TRUE for assuring the nearest neighbor of a point will not be itself.

Usage

nearest_neighbor_matrix(
  distances,
  subsetPoints,
  subsetNeighbors,
  contiguity,
  inner
)

Arguments

distances

distances between points and neighbors. A row is for a point and a column a neighbor.

subsetPoints

A vector of strictly positive integers giving the position of the points the nearest neighbor must be found. Positions starts from 1 like in R. Must be at least of length one and it would be better if values are unique.

subsetNeighbors

A vector of strictly positive integers giving the position of the elements of the set that can be consider has a neighbor for each point of subsetPoints. Positions starts from 1 like in R. Must be at least of length one and it would be better if values are unique.

contiguity

In the case there are contiguity constraints between points. If an element of subsetPoints and one of subsetNeighbors are not contiguous then the second element cannot be a neighbor of the first one. If no contiguity constraints must be considered a matrix with 0 rows and 0 columns must be sent. Otherwise must be a logical matrix with same number of rows and columns than distante.

Value

a vector of length p with for each point the position of its nearest neighbor (0 if it doesn't exist). Positions starts from 1 like in R.

Functions

nearest_neighbor_matrix(): case when distances is a matrix